factorize footer buttons
